refactor(normalize): relocate imported-HCL fixup engine to shared pkg
Move NormalizeImportedHCL and the resourceTypeFixups registry (all ~24 AWS
+ 1 GCP resource-type fixups, plus their literal/lifecycle helpers) out of
cmd/insideout-import/genconfig and into a new leaf package
pkg/composer/imported/normalize. genconfig previously owned this engine, so
pkg/composer could not run it over its own emitted /imported.tf without a
pkg -> cmd dependency inversion.

The new package imports only pkg/composer/imported (no cycle: imported
imports neither composer nor normalize), so both consumers can share the one
registry: the reverse-import pipeline (pkg/reverseimport/run.go) and the
composer path (a follow-up commit).

- normalize.go: the moved fixups.go, verbatim behavior. Exports
  NormalizeImportedHCL, ApplyResourceTypeFixupsReport/FixupResult (for
  genconfig's generated.tf progress logging), and MergeIgnoreChanges /
  IgnoreChangesTokens (the canonical lifecycle ignore_changes shape, moved
  from genconfig/cleanup.go so the fixups and cleanGenerated share one copy
  instead of two drifting ones). stringLitFromAttr is a tiny private copy of
  the genconfig import-ID helper.
- genconfig/cleanup.go, genconfig.go: delegate to the shared package.
- pkg/reverseimport/run.go: call normalize.NormalizeImportedHCL directly.

Pure refactor: no behavior change. The full fixups_test.go moves with the
engine (127 tests) and stays green; all genconfig and reverse-import tests
stay green.
