+++ pkg/composer/imported/normalize/normalize_test.go
@@ -1,4 +1,4 @@
-package genconfig
+package normalize
 
 import (
 	"reflect"
